requirements.txt
tensorflow==2.5.3
opencv-python==4.4.0.42
matplotlib==3.4.3
pyyaml==5.3.1
lmdb==1.3.0
scipy==1.4.1
scikit-learn==0.23.2
scikit-image==0.19.2
pandas==1.1.1